getById implementation

+++ backend/src/main/kotlin/outcobra/server/model/dto/InstitutionDto.kt
@@ -3,4 +3,4 @@ package outcobra.server.model.dto
 /**
  * Created by Florian on 29.10.2016.
  */
-data class InstitutionDto(val institutionId: Int, val institutionName : String)
+data class InstitutionDto(val institutionId: Long, val institutionName : String)
